requirements.txt
pydantic-settings
pydantic
python-dotenv
requests